pyproject.toml
[project]
name = "dbs-on-aws-webinar-demo"
version = "0.1.0"
description = "Databricks Asset Bundle for AWS Webinar demo"
requires-python = ">=3.10"

dependencies = [
    "databricks-sdk>=0.89.0",
    "databricks-agents",
    "databricks-langchain",
    "mlflow",
    "langchain",
    "langgraph",
    "psycopg2-binary",
]
